modified:   Assets/Fonts/fusion-pixel-12px SDF.asset modified:   Assets/Scripts/Gameplay/Entities/Creature.cs modified:   Assets/Scripts/Gameplay/Entities/Drop.cs modified:   Assets/Scripts/Gameplay/Entities/Player/Player.cs modified:   Assets/Scripts/Gameplay/Entities/Player/PlayerUI.cs modified:   Assets/Scripts/Gameplay/GM.cs modified:   Assets/Scripts/Gameplay/Map/Block.cs modified:   Assets/Scripts/Gameplay/Map/Chunk.cs modified:   Assets/Scripts/Gameplay/Map/Map.cs modified:   Assets/Scripts/Gameplay/Region.cs modified:   Assets/Scripts/Mods/ModLoading.cs modified:   Assets/Scripts/Network/NetworkMessages.cs modified:   Assets/Scripts/UI/InternalUIAdder.cs modified:   Assets/Scripts/UI/ParallaxBackground.cs

+++ Assets/Scripts/Gameplay/Entities/Player/PlayerUI.cs
@@ -1089,7 +1089,7 @@ namespace GameCore
                     {
                         rebornButton.button.interactable = false;
                         rebornTimerText.gameObject.SetActive(true);
-                        rebornTimerText.text.text = (player.deathTimer - Tools.time).ToString();
+                        rebornTimerText.text.text = ((int)(player.deathTimer - Tools.time)).ToString();
                     }
                 };
             }
